+++ Tests/CohesionKitTests/IdentityMapTests.swift
@@ -8,7 +8,7 @@ class IdentityMapTests: XCTestCase {
     func test_getForId_entityWasAdded_itReturnEntity() {
         let map = IdentityMap()
 
-        _ = map.update(Entity.hello)
+        _ = map.store(Entity.hello)
 
         XCTAssertNotNil(map.get(for: Entity.self, id: Entity.hello.id))
     }
@@ -17,7 +17,7 @@ class IdentityMapTests: XCTestCase {
         let map = IdentityMap()
         var cancellables: Set<AnyCancellable> = []
 
-        map.update(Entity.hello)
+        map.store(Entity.hello)
             .sink(receiveValue: { _ in })
             .store(in: &cancellables)
 
@@ -26,10 +26,10 @@ class IdentityMapTests: XCTestCase {
         XCTAssertNil(map.get(for: Entity.self, id: Entity.hello.id))
     }
 
-    func test_updateIfPresent_valueIsNotPresent_itReturnNil() {
+    func test_storeIfPresent_valueIsNotPresent_itReturnNil() {
         let map = IdentityMap()
 
-        XCTAssertNil(map.updateIfPresent(Entity.hello))
+        XCTAssertNil(map.storeIfPresent(Entity.hello))
     }
 
     func test_publisherForId_entityIsAddedAfterRequesting_entityIsEmitted() {
@@ -38,7 +38,7 @@ class IdentityMapTests: XCTestCase {
         var receivedValue: Entity?
         var cancellables: Set<AnyCancellable> = []
 
-        _ = map.update(Entity.hello)
+        _ = map.store(Entity.hello)
 
         publisher
             .sink(receiveValue: { receivedValue = $0 })